Refactor services with interfaces and implementations
