+++ UniVerseAPI/universe/src/main/java/com/unidev/universe/services/EventService.java
@@ -1,110 +1,25 @@
 package com.unidev.universe.services;
 
-import com.unidev.universe.entities.User;
-import com.unidev.universe.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import com.unidev.universe.repository.EventRepository;
 import com.unidev.universe.entities.GroupEvent;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
 
-@Service
-@RequiredArgsConstructor
-public class EventService {
+public interface EventService {
+    List<GroupEvent> getAllEvents();
 
-    private final EventRepository eventRepository;
-    private final UserRepository userRepository;
+    List<GroupEvent> getTrendingEvents();
 
-    public List<GroupEvent> getAllEvents() {
-        return eventRepository.findAll();
-    }
+    GroupEvent getEventById(Long eventId);
 
-    public List<GroupEvent> getTrendingEvents() {
-        List<GroupEvent> events = eventRepository.findAll();
-        List<GroupEvent> trendingEvents = new ArrayList<>();
+    boolean isAttending(Long eventId, String username);
 
-        for (GroupEvent event: events) {
-            if(event.getAttendees().size() > 10){
-                trendingEvents.add(event);
-            }
-        }
+    void attendEvent(Long eventId, String username);
 
-        return trendingEvents;
-    }
+    void removeAttending(Long eventId, String username);
 
-    public GroupEvent getEventById(Long eventId) {
-        Optional<GroupEvent> optionalEvent = eventRepository.findById(eventId);
-        return optionalEvent.orElse(null);
-    }
+    GroupEvent createEvent(GroupEvent event);
 
-    public boolean isAttending(Long eventId, String username){
-        Optional<GroupEvent> optionalEvent = eventRepository.findById(eventId);
-        User user = userRepository.findByUsername(username);
+    GroupEvent updateEvent(Long eventId, GroupEvent updatedEvent);
 
-        if (user != null && optionalEvent.isPresent()) {
-            GroupEvent event = optionalEvent.get();
-
-            for (User attendee: event.getAttendees()) {
-                if(Objects.equals(attendee, user)){
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    public void attendEvent(Long eventId, String username) {
-        User user = userRepository.findByUsername(username);
-        Optional<GroupEvent> optionalEvent = eventRepository.findById(eventId);
-
-        if(user != null && optionalEvent.isPresent()){
-            GroupEvent event = optionalEvent.get();
-            event.getAttendees().add(user);
-
-            eventRepository.save(event);
-        }
-    }
-
-    public void removeAttending(Long eventId, String username) {
-        User user = userRepository.findByUsername(username);
-        Optional<GroupEvent> optionalEvent = eventRepository.findById(eventId);
-
-        if(user != null && optionalEvent.isPresent()){
-            GroupEvent event = optionalEvent.get();
-            event.getAttendees().remove(user);
-
-            eventRepository.save(event);
-        }
-    }
-
-    //TODO: Authorization with logged users, only specific users can create, edit and delete events!
-    public GroupEvent createEvent(GroupEvent event) {
-        return eventRepository.save(event);
-    }
-
-    //TODO: Authorization with logged users, only specific users can create, edit and delete events!
-    public GroupEvent updateEvent(Long eventId, GroupEvent updatedEvent) {
-        Optional<GroupEvent> optionalEvent = eventRepository.findById(eventId);
-
-        if (optionalEvent.isPresent()) {
-            GroupEvent existingEvent = optionalEvent.get();
-            existingEvent.setTitle(updatedEvent.getTitle());
-            existingEvent.setDescription(updatedEvent.getDescription());
-            existingEvent.setDate(updatedEvent.getDate());
-            return eventRepository.save(existingEvent);
-        } else {
-            return null; 
-        }
-    }
-
-    //TODO: Authorization with logged users, only specific users can create, edit and delete events!
-    public void deleteEvent(Long eventId) {
-        eventRepository.deleteById(eventId);
-    }
+    void deleteEvent(Long eventId);
 }
