+++ UniVerseAPI/universe/src/main/java/com/unidev/universe/services/NotificationService.java
@@ -1,42 +1,13 @@
 package com.unidev.universe.services;
 
 import com.unidev.universe.entities.Notification;
-import com.unidev.universe.entities.User;
-import com.unidev.universe.repository.NotificationRepository;
-import com.unidev.universe.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-public class NotificationService {
-    private final NotificationRepository notificationRepository;
-    private final UserRepository userRepository;
+public interface NotificationService {
+    void addNotification(Notification notification);
 
-    public void addNotification(Notification notification){
-        User user = userRepository.findByUsername(notification.getRecipientName());
-        notification.setUser(user);
+    List<Notification> getUserNotifications(String username);
 
-        notificationRepository.save(notification);
-    }
-
-    public List<Notification> getUserNotifications(String username){
-        User user = userRepository.findByUsername(username);
-
-        return notificationRepository.findAllByUser(user);
-    }
-
-    public void readNotifications(String username){
-        User user = userRepository.findByUsername(username);
-
-        List<Notification> userNotifications = notificationRepository.findAllByUser(user);
-
-        for (Notification notification: userNotifications) {
-            notification.setRead(true);
-        }
-
-        notificationRepository.saveAll(userNotifications);
-    }
+    void readNotifications(String username);
 }
